+++ lib/llm/active_record/acts_as_llm.rb
@@ -113,6 +113,13 @@ module LLM::ActiveRecord
       end
 
       ##
+      # @see LLM::Context#functions?
+      # @return [Boolean]
+      def functions?
+        ctx.functions?
+      end
+
+      ##
       # @see LLM::Context#returns
       # @return [Array<LLM::Function::Return>]
       def returns